Bugfix:Android - Fix case where UltravioletContext property is accessed before context is created.
If the app starts while the device is not in the default orientation then Android will invoke the OnConfigurationChanged method. This can happen before the the UltravioletContext is created. Accessing UltravioletContext in this case will trigger the assertion which checks that the context is created. To work around this, check that the create flag is true before accessing the context.