Introduced UltravioletApplicationAdapter class in the Ultraviolet project. This class is intended to be extended to implement application/game logic. Game classes should no longer extend UltravioletApplication as classes derived from this cannot be used in a csproj and shared between NETCore and Android without preprocessor magic. Changed UvDebug from a Shared (.shprog) project to a class library (.csproj).
Added an ApplicationAdapter implementation for tests.

UltravioletApplication in the Android shim will be renamed to UltravioletAndroidApplicationHost in a follow up commit.
UltravioletApplication in the NETCore shim will be renamed to NETCoreApplicationHost name as well.
UltravioletApplicationAdapter will possibly be renamed to UltravioletApplication.

The end result is that entry point projects will be treated as hosts for the "applications".

Reference Ultraviolet.SDL in platform shim projects.
Create context config and context in base application class.
Give both application and application adapter a chance to configure the context.

Updated tests to use application adapter.